pyproject.toml
[tool.poetry]
name = "youtube-shorts-gen"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
ffmpeg-python = "^0.2.0"
pillow = "^10.4.0"


[tool.poetry.group.dev.dependencies]
ipython = "^8.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
